app/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
fastapi==0.104.1
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
motor==3.3.2
numpy==1.26.2
opencv-python==4.8.1.78
pydantic==2.5.2
pydantic_core==2.14.5
pymongo==4.6.0
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
